Refactor utils.notifications.plex_update to plugins.plex.plex_update
