+++ core/utils/__init__.py
@@ -19,7 +19,6 @@ from core.utils.identification import category_search, find_imdbid
 from core.utils.links import copy_link, replace_links
 from core.utils.naming import clean_file_name, is_sample, sanitize_name
 from core.utils.network import find_download, server_responding, test_connection, wake_on_lan, wake_up
-from core.utils.notifications import plex_update
 from core.utils.parsers import (
     parse_args,
     parse_deluge,
